workflows: Init pre-commit checks workflow
Implement a simple workflow that starts pre-commit checks based on the
implemented configuration file in the root directory of the repository.
The workflow initially should be trigerred manually only.

The workflow prepares minimalist environment on Ubuntu 24.04 LTS
operating system and runs checks for the whole repository.

At the end the produced log file is pushed as an artifact of the workflow
run to be able to debug offline.

For the correct behaviour of the code mixed-line-endings rule for .hex files
should be excluded (they should have CRLF line endings) as well as
Kconfig files should be based on tabs instead of spaces for the main
indentation.