devcontainer, workflows: Organize scripts
Due to growing number of helper Bash scripts, to not mislead users of the
repository, their location has been changed.

The scripts should not be treated as a part of the actual
implementation, but as helper tools related to DevContainer or Github
Action workflows (in this case).

Thus, scripts have been moved to newly created directories in respective
paths.