workflows: Implement release repository workflow
The workflow to help with releasing process has been implemented.

It is automatically configuring the Git user, creates/pushes tag,
prepares archive and at the end creates the release.

The Git user is configured based on the actor that starts the workflow
and sets the e-mail accordingly to noreply variant.

The tag is created based on the provided input name. It is a lightweight
variant.

The repository is packed as an *.zip archive with default compression
level. It excludes .git/ directory (only sources are being processed).

The release is created based on the provided in the initial form values.
The title is set equally to the name of the tag.