runner, workflows: Fix handling of getops arguments
In some of the scripts rare error could occur. The cause is explained
below.

If someone calls a script and forgets to provide a value for the option
which needs an argument (e.g. -x), but additionally use other option
right after (e.g. -y) -> ./script -x -y test, it leads to an incorrect
behavior.
This causes that the second option (-y) is assigned as a string value to
the first one (-x) instead of reporting that the value is missing for the
first option.
It might be a bug inside getopts. As an outcome, let's prevent this event.

The commit provide simple if condition preventing failure of the script due to
the above event.